backend/speech/requirements.txt
fastapi
uvicorn[standard]
numpy
librosa
soundfile
pydub
tensorflow-cpu==2.15.0
